feat: 수기 입력 소수점 첫째 자리 지원
- 입력 검증 정규식 정수 → 소수점 1자리 허용
- parseInt → parseFloat 변환으로 소수점 값 처리
- 백엔드 스키마 int → float으로 변경
- 기존 데이터 불러올 때 소수점 표시

+++ backend/app/models/schemas.py
@@ -109,9 +109,9 @@ class SharedAnalysisSetResponse(BaseModel):
 # ── Story 5.1: 비상장사 수기 입력 ────────────────────────
 class ManualFinancialEntry(BaseModel):
     bsns_year: str = Field(..., pattern=r'^\d{4}$')
-    revenue: Optional[int] = None
-    operating_profit: Optional[int] = None
-    net_income: Optional[int] = None
+    revenue: Optional[float] = None
+    operating_profit: Optional[float] = None
+    net_income: Optional[float] = None
 
 
 class ManualCompanyCreate(BaseModel):
